+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 7.0.0a4 (unreleased)
 --------------------
 
+- remove layer computation from controlpanel template
+  [erral]
+
 - editorconfig: Don't use inline comments.
   Inline comments are invalid since editorconfig v0.15.0.
   [thet]